Create and Extend CountTask Abstract Class
