+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowersCountTask.java
@@ -9,29 +9,15 @@ import edu.byu.cs.tweeter.model.domain.User;
 /**
  * Background task that queries how many followers a user has.
  */
-public class GetFollowersCountTask extends AuthenticatedTask {
+public class GetFollowersCountTask extends CountTask {
     private static final String LOG_TAG = "GetFollowersCountTask";
 
-    public static final String COUNT_KEY = "count";
-
-    /**
-     * The user whose follower count is being retrieved.
-     * (This can be any user, not just the currently logged-in user.)
-     */
-    private User targetUser;
-
     public GetFollowersCountTask(AuthToken authToken, User targetUser, Handler messageHandler) {
-        super(messageHandler, authToken);
-        this.targetUser = targetUser;
+        super(messageHandler, authToken, targetUser);
     }
 
     @Override
     protected void executeTask() {
         // TODO: Nothing to override currently
-    }
-
-    @Override
-    protected void loadSuccessBundle(Bundle msgBundle) {
-        msgBundle.putInt(COUNT_KEY, 20);
     }
 }
