+++ app/src/main/java/edu/byu/cs/tweeter/client/model/service/backgroundTask/GetFollowingCountTask.java
@@ -9,29 +9,15 @@ import edu.byu.cs.tweeter.model.domain.User;
 /**
  * Background task that queries how many other users a specified user is following.
  */
-public class GetFollowingCountTask extends AuthenticatedTask {
+public class GetFollowingCountTask extends CountTask {
     private static final String LOG_TAG = "GetFollowingCountTask";
 
-    public static final String COUNT_KEY = "count";
-
-    /**
-     * The user whose following count is being retrieved.
-     * (This can be any user, not just the currently logged-in user.)
-     */
-    private User targetUser;
-
     public GetFollowingCountTask(AuthToken authToken, User targetUser, Handler messageHandler) {
-        super(messageHandler, authToken);
-        this.targetUser = targetUser;
+        super(messageHandler, authToken, targetUser);
     }
 
     @Override
     protected void executeTask() {
         // TODO: Nothing to override currently
-    }
-
-    @Override
-    protected void loadSuccessBundle(Bundle msgBundle) {
-        msgBundle.putInt(COUNT_KEY, 20);
     }
 }
